/add Manejo de excepciones
- Al momento de crear una mascota con un cliente no existente
- Crear un cliente con una cédula ya existente
- Modificar la cédula de un cliente por otra ya existente

+++ demo/src/main/java/com/example/demo/service/ClienteService.java
@@ -5,13 +5,17 @@ import java.util.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.example.demo.controller.ClientExistingException;
+import com.example.demo.controller.ClientUpdatingException;
 import com.example.demo.controller.NotClientFoundException;
+import com.example.demo.controller.NotPetFoundException;
+
 import com.example.demo.entity.Cliente;
 import com.example.demo.entity.Mascota;
+
 import com.example.demo.repository.ClienteRepository;
 import com.example.demo.repository.MascotaRepository;
 
-import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 
 @Service
@@ -29,7 +33,7 @@ public class ClienteService {
         // Si la mascota ya existe en la base de datos, solo actualiza el vínculo
         if (mascota.getId() != null) {
             Mascota existingMascota = mascotaRepository.findById(mascota.getId())
-                .orElseThrow(() -> new EntityNotFoundException("Mascota no encontrada con ID: "));
+                .orElseThrow(() -> new NotPetFoundException(0L));
             existingMascota.setNombre(mascota.getNombre());
             existingMascota.setRaza(mascota.getRaza());
             existingMascota.setEdad(mascota.getEdad());
@@ -85,7 +89,11 @@ public class ClienteService {
     }
 
     public void add(Cliente cliente){
-        clienteRepository.save(cliente);
+        if(clienteRepository.findByCedula(cliente.getCedula()) != null){
+            throw new ClientExistingException(cliente.getCedula());
+        } else{
+            clienteRepository.save(cliente);
+        }
     }
 
     // public void editarCliente(Cliente cliente){
@@ -93,8 +101,12 @@ public class ClienteService {
     // }
 
     public void update(Cliente cliente){
-        cliente.setId(clienteRepository.findByCedula(cliente.getCedula()).getId());
-        clienteRepository.deleteById(cliente.getCedula());
-        clienteRepository.save(cliente);
+        if(clienteRepository.existsByCedula(cliente.getCedula())){
+            throw new ClientUpdatingException(cliente.getCedula());
+        } else{ 
+            cliente.setId(clienteRepository.findByCedula(cliente.getCedula()).getId());
+            clienteRepository.deleteById(cliente.getCedula());
+            clienteRepository.save(cliente);
+        }
     }
 }
